Uploading last year's code
Square 1 for this year, this should be flight ready for a system containing an IMU (BNO055), Altimeter (BMP388), and Multiplexer (TCA9548A). This has most of the code needed for an Adafruit V3 GPS as well, but it's using a UART1 pin, which have been difficult to get working properly.